requirements.txt
alembic==1.3.1
appnope==0.1.0
backcall==0.1.0
blinker==1.4
Click==7.0
decorator==4.4.2
dominate==2.5.1
elasticsearch==7.6.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
flask-shell-ipython==0.4.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
six==1.13.0
SQLAlchemy==1.3.11
traitlets==4.3.3
urllib3==1.25.9
visitor==0.1.3
wcwidth==0.1.9
Werkzeug==0.16.0
WTForms==2.2.1
